Introducing feature: Perk Achievements
This feature branch tracks user achievements in a more Fallout(tm) like
fashion.  Each daily goal is awarded a new perk.  Achieving all daily
goals earns a final perk "Strength".

As new perks are achieved, the vaulBoy icon is updated with that perk
image.

This code lays the groundwork for new perks to easily be added.