Add feature Perk Card notification
This feature includes a display of achievement in the form of a perk
card.  The card will fade into the screen to notify the user what
achievement was unlocked.  The window will dissapear after it is
clicked.

Once a perk is notified to the user it will never be notified again.
When perks are reloaded the next day the notification resets.