Fixed akward opacity animation for perk cards
From time to time a perk card will animate roughly and flash the icon
instead of smoothly transitioning the image on.

This fix changes the way we show the card.  We no longer turn on the
opacity of the perk card.  Now we cover the card and turn off the
opacity of the cover.

This results in a much smoother and consistent transition.